customer get restaurants by location current code

+++ Backend/DataAccess/Models/Restaurant.cs
@@ -21,7 +21,7 @@ public class Restaurant
     public string? WorkingHours { get; set; }
 
     [Column(TypeName = "decimal(5,2)")]
-    public float? TotalScore { get; set; }
+    public decimal? TotalScore { get; set; }
 
     //Relationships
